requirements.txt
asgiref==3.11.0
channels==4.3.2
dj-database-url==3.0.1
Django==4.2.27
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0

#djangorestframework==4.14.0
django-cors-headers==4.3.1
#djangorestframework-simplejwt==5.3.0
#channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
pillow==10.1.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
